Deduplicate the `getArchitectureFromInfo` function
