+++ extensions/positron-python/src/client/pythonEnvironments/info/interpreter.ts
@@ -7,33 +7,12 @@ import {
     interpreterInfo as getInterpreterInfoCommand,
     InterpreterInfoJson,
 } from '../../common/process/internal/scripts';
+// --- Start Positron ---
+import { getArchitectureFromInfo } from '../../common/process/internal/scripts/architecture';
+// --- End Positron ---
 import { ShellExecFunc } from '../../common/process/types';
 import { replaceAll } from '../../common/stringUtils';
-import { Architecture } from '../../common/utils/platform';
 import { copyPythonExecInfo, PythonExecInfo } from '../exec';
-
-// --- Start Positron ---
-/**
- * Get the Architecture enum value from interpreter info.
- * Prefers the explicit architecture field, falls back to is64Bit for compatibility.
- */
-function getArchitectureFromInfo(raw: InterpreterInfoJson): Architecture {
-    if (raw.architecture) {
-        const normalized = raw.architecture.toLowerCase();
-        if (normalized === 'arm64' || normalized === 'aarch64') {
-            return Architecture.arm64;
-        }
-        if (normalized === 'x86_64' || normalized === 'amd64' || normalized === 'x64') {
-            return Architecture.x64;
-        }
-        if (normalized === 'x86' || normalized === 'i386' || normalized === 'i686') {
-            return Architecture.x86;
-        }
-    }
-    // Fall back to is64Bit for backward compatibility
-    return raw.is64Bit ? Architecture.x64 : Architecture.x86;
-}
-// --- End Positron ---
 
 /**
  * Compose full interpreter information based on the given data.
